+++ PermuteMMO.Lib/Util/JsonDecoder.cs
@@ -12,7 +12,7 @@ public static class JsonDecoder
     /// <summary>
     /// Wrapper to deserialize the json using whatever package this project is currently using.
     /// </summary>
-    public static T Deserialize<T>(string path) where T : class => JsonConvert.DeserializeObject<T>(path);
+    public static T Deserialize<T>(string json) where T : class => JsonConvert.DeserializeObject<T>(json);
 
     /// <summary>
     /// Converts the json string back to a usable dictionary.
